Update tutorial with HTTP service targets
Previously, the tutorial showed how a service could be implemented
with a device module target.

In a recent change to TDM however, support for HTTP targets was added,
which is recommended over device modules.

This change updates the tutorial to show how a service can be
implemented with HTTP targets instead.

Additional changes:
- '<dev_perform>' changes to '<invoke_service_action>'
in the plan for the 'call' action.

Taiga story #1122
Taiga task #1234

Change-Id: Icc780f23c1611d35dd1a0c4063e1c305754c82a6