Docs: Add links to DDDs
Previously, the 'examples' page in the docs only contained a link to the
Github repository, not to specific DDD examples themselves. For a reader
of the examples page to actually find the intended DDD example on Github
he or she needed to do some non-trivial detective work.

This change replaces the repository link with links to the DDDs
themselves for each example in the docs, no longer requiring detective
work from readers.

Change-Id: Idb76b678551730aa80403f01607616ad7a516468