Docs: Add 'semantic inputs' to frontend API
This change adds documentation for 'semantic inputs' to the HTTP API for
frontends.

Additionally, this change:
- converts the version number of the same API from a float to a string
- renames field 'input' to 'natural_language_input'
- adds a note about the 'warnings' field of the response format
- clarifies how requests can be combined
- adds information about the 'session' object in responses
- clarifies how the response looks when an error is encountered
- converts requests from:

{
  "request": {
    "type": "<name>",
    ...
  }
}

to

{
  "request": {
    "<name>": {...}
  }
}

This way, a request no longer has a single type, instead allowing more
than one type of object to be included in the request simultaneously.
This is useful if e.g. NLU is done externally, so that TDM can take both
'natural_language_input' and 'semantic_input' into account in the same
request. However, in this initial version, the only combinable request
is 'start_session'. Other combinations take more work but may be
implemented in the future.

Taiga epic #1631

Change-Id: I50fc8265057d39f4c5b600cab1b5c508881d74af