Docs: Change version to 'master' in links
Before this change, links from the docs to the source code was done to
an explicit version.

Recently however, TDM went from being semantically versioned, to
automated releases after every change.

This change updates the version in the links to 'master', also
eliminating the need for future changes to the links.

Change-Id: I6fd15fc9c81fe5f9f84edd39a038dd91fc1c6fe7