Docs: Move shared objects to a separate document
This change takes some of the objects sections and examples from
'http_frontend.md' to their own new document called
'shared_objects.md'. This is done so, because these objects will be
used in requests and responses in more than one API and that way they
can refer to them in the same one place.

Change-Id: Ieb8a3d3234606499a4a0afcd440ff719cba7636e