jenkinsfile-publish-docs: Mount /etc/passwd
Before this change, when trying to publish docs by pushing to the Github
repo over SSH, the SSH client inside the container complained with

  No user exists for uid 1000

since the uid was taken from the host, which didn't exist in the Docker
container set up temporarily by the pipeline.

A solution is to mount /etc/passwd of the host into the container, which
means that the user appears to exist, and the error does not occur. The
solution has been reported for instance here:

  https://stackoverflow.com/questions/42404473/docker-plugin-for-
jenkins-pipeline-no-user-exists-for-uid-1005

Change-Id: I7dd92eb865a1b6926805216b5309e777ee2ecf7c