jenkinsfile-publish-docs: Accept unknown host
Before this change, when pushing the built docs to Github, the git push
command failed because the github.com SSH signature didn't exist in
~/.ssh/known_hosts.

This change instructs ssh to accept unknown hosts, avoiding the problem.

Additionally, this change instructs ssh to write the unknown host to
/dev/null instead of the default ~/.ssh/known_hosts to avoid missing
write permission inside the container. This is needed since there is no
home folder for the user's UID, which actually only exists on the host,
not in the container, anyway.

Change-Id: Idc94c80280153c52d22a6085757cd87f5110ab04