Change http_service urls
Since the http-service container was separated into a different
deployment, the URL in the service interface cannot point to a local
served port anymore, but to the name of the HTTP service
deployment.

This changes the documentation for DDD developers to reflect the
new urls necessary for the HTTP service running on a deployment.

Change-Id: I9445967e8f74cb3ece941a3251a03cd470125c0f